Add initial DB seed

+++ db/migrate/20230815172022_create_users.rb
@@ -4,10 +4,10 @@ class CreateUsers < ActiveRecord::Migration[6.1]
       t.string :name, null: false
       t.bigint :discord_id
       t.string :old_discord_name
-      t.integer :current_points, default: 0
-      t.integer :redeemed_points, default: 0
-      t.integer :earned_points, default: 0
-      t.integer :premium_points, default: 0
+      t.float :current_points, default: 0
+      t.float :redeemed_points, default: 0
+      t.float :earned_points, default: 0
+      t.float :premium_points, default: 0
       t.timestamps
     end
   end
